Removed duplicate format_value_string definition
